+++ src/core/builtins/builtins_request.ml
@@ -180,7 +180,14 @@ let _ =
 
 let _ =
   Lang.add_builtin ~base:request "duration" ~category:`Liquidsoap
-    [("", Lang.string_t, None, None)]
+    [
+      ( "metadata",
+        Lang.metadata_t,
+        Some (Lang.list []),
+        Some "Optional metadata used to decode the file, e.g. `ffmpeg_options`."
+      );
+      ("", Lang.string_t, None, None);
+    ]
     (Lang.nullable_t Lang.float_t)
     ~descr:
       "Compute the duration in seconds of audio data contained in a request. \
@@ -188,7 +195,8 @@ let _ =
        typically if the file was not recognized as valid audio."
     (fun p ->
       let f = Lang.to_string (List.assoc "" p) in
-      try Lang.float (Request.duration f) with _ -> Lang.null)
+      let metadata = Lang.to_metadata (List.assoc "metadata" p) in
+      try Lang.float (Request.duration ~metadata f) with _ -> Lang.null)
 
 let _ =
   Lang.add_builtin ~base:request "id" ~category:`Liquidsoap