+++ src/core/decoder/ogg_metadata_plug.ml
@@ -39,7 +39,7 @@ let file_extensions =
     "File extensions used for decoding metadata using native ogg parser."
     ~d:["ogg"]
 
-let get_tags parse fname =
+let get_tags ~metadata:_ parse fname =
   try
     if
       not